+++ src/common/components/Icon/index.js
@@ -3,16 +3,26 @@ import styles from './styles.less';
 import classNames from 'classnames/bind';
 const cx = classNames.bind(styles);
 
-const Icon = ({ name, size, className, src, button, ...props }) => (
-  <div
-    {...props}
-    className={cx('container', size, className, button && 'button')}
-    style={{
-      /* eslint prefer-template:0 */
-      backgroundImage: (src || name) && 'url(' + (src || require(`assets/images/${name}`)) + ')',
-    }}
-  />
-);
+const Icon = ({ name, size, className, src, button, ...props }) => {
+  let image;
+
+  try {
+    image = require(`assets/images/${name}`);
+  } catch (ex) {
+    image = undefined;
+  }
+
+  return (
+    <div
+      {...props}
+      className={cx('container', size, className, button && 'button')}
+      style={{
+        /* eslint prefer-template:0 */
+        backgroundImage: (src || name) && 'url(' + (src || image) + ')',
+      }}
+    />
+  );
+};
 
 Icon.defaultProps = {
   size: 'mini',