+++ src/common/layouts/Summary/index.js
@@ -2,10 +2,13 @@ import { PropTypes } from 'react';
 import cx from 'classnames';
 
 import styles from './styles.less';
+import Icon from 'common/components/Icon';
 
 const Summary = ({ className, leftIcon, title, subTitle }) => (
   <div className={cx(className, styles.root)}>
-    <div className={styles.icon}>{leftIcon}</div>
+    <div className={styles.icon}>
+      <Icon {...leftIcon} />
+    </div>
 
     <div className={styles.content}>
       <span className={styles.title}>{title}</span>
@@ -16,7 +19,7 @@ const Summary = ({ className, leftIcon, title, subTitle }) => (
 
 Summary.propTypes = {
   className: PropTypes.string,
-  leftIcon: PropTypes.any,
+  leftIcon: PropTypes.object,
   title: PropTypes.any,
   subTitle: PropTypes.any,
 };