+++ src/features/User/components/Fractal/index.js
@@ -1,23 +1,27 @@
 import { PropTypes } from 'react';
 
 import Summary from 'common/layouts/Summary';
-import Icon from 'common/components/Icon';
 import ProgressBar from 'common/components/ProgressBar';
+import Redacted from 'common/components/Redacted';
 
-const Fractal = ({ level }) => (
-  <Summary
-    leftIcon={<Icon name="raid.png" size="xlarge" />}
-    title="Fractal Level"
-    subTitle={
-      <ProgressBar
-        current={level || 0}
-        max={100}
-        backgroundColor="rgb(41, 41, 41)"
-        barColor="rgb(85, 35, 164)"
-      />
-    }
-  />
-);
+const Fractal = ({ level }) => {
+  const redact = !level;
+
+  return (
+    <Summary
+      leftIcon={{ name: 'raid.png', size: 'xlarge' }}
+      title={<Redacted redact={redact}>Fractal Level</Redacted>}
+      subTitle={
+        <ProgressBar
+          current={level || 0}
+          max={100}
+          backgroundColor="rgb(41, 41, 41)"
+          barColor="rgb(85, 35, 164)"
+        />
+      }
+    />
+  );
+};
 
 Fractal.propTypes = {
   level: PropTypes.number,