+++ src/features/User/components/RaidSummary/index.js
@@ -1,8 +1,8 @@
 import { PropTypes } from 'react';
 
 import ProgressBar from 'common/components/ProgressBar';
-import Icon from 'common/components/Icon';
 import Summary from 'common/layouts/Summary';
+import Redacted from 'common/components/Redacted';
 
 const RAID_KILL_ACHIEVEMENTS = {
   2654: 'Vg', // Vale Guardian
@@ -17,11 +17,12 @@ const RAID_KILL_ACHIEVEMENTS = {
 const TOTAL_KILLS = Object.keys(RAID_KILL_ACHIEVEMENTS).length;
 
 const RaidSummary = ({ userAchievements, className, simple }) => {
+  const redact = !userAchievements.length;
   const achievements = userAchievements.filter(
     ({ id, done }) => RAID_KILL_ACHIEVEMENTS[id] && done
   );
 
-  const raidKills = `Raid Boss Kills ${achievements.map(({ id }) =>
+  const raidKills = `Raid Bosses ${achievements.map(({ id }) =>
     `[${RAID_KILL_ACHIEVEMENTS[id]}]`).join('')}`;
 
   if (simple) {
@@ -31,8 +32,8 @@ const RaidSummary = ({ userAchievements, className, simple }) => {
   return (
     <Summary
       className={className}
-      leftIcon={<Icon name="raid.png" size="xlarge" />}
-      title={raidKills}
+      leftIcon={{ name: 'raid.png', size: 'xlarge' }}
+      title={<Redacted redact={redact}>{raidKills}</Redacted>}
       subTitle={
         <ProgressBar
           current={achievements.length}